get-job/requirements.txt
# Function dependencies, for example:
# package>=version
firebase-admin
google-cloud-firestore
matplotlib
qiskit